+++ src/components/Navbar/style.js
@@ -1,0 +1,23 @@
+import { makeStyles } from "@material-ui/core";
+
+const useStyles = makeStyles((theme) => ({
+  appBar: {
+    zIndex: theme.zIndex.drawer + 1,
+  },
+  appBarShift: {
+    marginLeft: theme.drawerWidth,
+    width: `calc(100% - ${theme.drawerWidth}px)`,
+  },
+  menuButton: {
+    marginRight: 36,
+  },
+  hide: {
+    display: "none",
+  },
+  content: {
+    flexGrow: 1,
+    padding: theme.spacing(3),
+  },
+}));
+
+export default useStyles;
